+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- prune: invoke a stop, before the destroy, even if the pot is not runing
+
+### Fixed
+- stop: make the pkill on ncat more robust
 
 ## [0.9.1] 2019-08-20
 ### Added